# apps/engine/src/__generated__/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stripe Sync Engine",
    "version": "1.0.0",
    "description": "Stripe Sync Engine — stateless, one-shot source/destination sync over HTTP.\nAll sync endpoints accept configuration via the `X-Pipeline` header (JSON-encoded PipelineConfig). Optional cursor state can be provided via `X-State`.\n\n## Endpoints\n\n| Method | Path | Summary |\n|--------|------|---------|\n| GET | /health | Health check |\n| POST | /pipeline_setup | Set up destination schema |\n| POST | /pipeline_teardown | Tear down destination schema |\n| GET | /pipeline_check | Check connector connection |\n| POST | /source_discover | Discover available streams |\n| POST | /pipeline_read | Read records from source |\n| POST | /pipeline_write | Write records to destination |\n| POST | /pipeline_sync | Run sync pipeline (read → write) |\n| GET | /meta/sources | List available source connectors |\n| GET | /meta/sources/{type} | Get source connector spec |\n| GET | /meta/destinations | List available destination connectors |\n| GET | /meta/destinations/{type} | Get destination connector spec |"
  },
  "paths": {
...
    "/pipeline_setup": {
      "post": {
        "operationId": "pipeline_setup",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Set up destination schema",
        "description": "Creates destination tables and applies migrations. Safe to call multiple times.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Setup complete",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_teardown": {
      "post": {
        "operationId": "pipeline_teardown",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Tear down destination schema",
        "description": "Drops destination tables. Irreversible.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Teardown complete"
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_check": {
      "get": {
        "operationId": "pipeline_check",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Check connector connection",
        "description": "Validates the source/destination config and tests connectivity.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection check result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "source": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "succeeded",
                            "failed"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "status"
                      ],
                      "additionalProperties": false
                    },
                    "destination": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "succeeded",
                            "failed"
                          ]
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "status"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "source",
                    "destination"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/source_discover": {
      "post": {
        "operationId": "source_discover",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Discover available streams",
        "description": "Streams NDJSON messages (catalog, logs, traces) for the configured source.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "NDJSON stream of discover messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_read": {
      "post": {
        "operationId": "pipeline_read",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Read records from source",
        "description": "Streams NDJSON messages (records, state, catalog). Optional NDJSON body provides live events as input.",
        "parameters": [
          {
            "in": "query",
            "name": "state_limit",
            "schema": {
              "description": "Stop streaming after N state messages.",
              "example": "100",
              "type": "integer",
              "exclusiveMinimum": 0,
              "maximum": 9007199254740991
            },
            "description": "Stop streaming after N state messages."
          },
          {
            "in": "query",
            "name": "time_limit",
            "schema": {
              "description": "Stop streaming after N seconds.",
              "example": "10",
              "type": "number",
              "exclusiveMinimum": 0
            },
            "description": "Stop streaming after N seconds."
          },
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          },
          {
            "in": "header",
            "name": "x-state",
            "required": false,
            "description": "JSON-encoded per-stream cursor state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              }
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/x-ndjson": {
              "schema": {
                "$ref": "#/components/schemas/SourceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NDJSON stream of sync messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/pipeline_write": {
      "post": {
        "operationId": "pipeline_write",
        "tags": [
          "Stateless Sync API"
        ],
        "summary": "Write records to destination",
        "description": "Reads NDJSON messages from the request body and writes them to the destination. Pipe /read output as input.",
        "parameters": [
          {
            "in": "header",
            "name": "x-pipeline",
            "required": true,
            "description": "JSON-encoded PipelineConfig",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineConfig"
                }
              }
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-ndjson": {
              "schema": {
                "$ref": "#/components/schemas/Message"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "NDJSON stream of write result messages",
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid params",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {}
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/meta/sources/{type}": {
      "get": {
        "operationId": "meta_sources_get",
        "tags": [
          "Meta"
        ],
        "summary": "Get source connector spec",
        "parameters": [
          {
            "in": "path",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Source connector spec",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "config_schema": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "config_schema"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Source connector not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiscoverOutput": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/CatalogMessage"
          },
          {
            "$ref": "#/components/schemas/LogMessage"
          },
          {
            "$ref": "#/components/schemas/TraceMessage"
          }
        ],
        "type": "object",
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "catalog": "#/components/schemas/CatalogMessage",
            "log": "#/components/schemas/LogMessage",
            "trace": "#/components/schemas/TraceMessage"
          }
        }
      },
...
      "SourceInput": {
        "oneOf": [
          {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "const": "stripe"
              },
              "stripe": {
                "$ref": "#/components/schemas/SourceStripeInput"
              }
            },
            "required": [
              "type",
              "stripe"
            ],
            "additionalProperties": false
          }
        ],
        "type": "object",
        "discriminator": {
          "propertyName": "type"
        }
      },
...
    }
  }
}